Front-end #7: Corrección CONFLICTOS debido a la modificacion de app.py del Front-end tras modularizar con Blueprints.
