+++ API/app/app.py
@@ -21,4 +21,4 @@ app.register_blueprint(reservas_bp, url_prefix='/reservas')
 app.register_blueprint(users_bp, url_prefix='/users')
 
 if __name__ == "__main__":
-    app.run(host='127.0.0.0.0', port=5000, debug=True) 
+    app.run(host='127.0.0.0', port=5000, debug=True)
