+++ Frontend/templates/404.html
@@ -38,7 +38,7 @@
             </ul>
             <ul class="navbar-nav">
                 <li class="nav-item">
-                    <a class="nav-link" href="{{url_for('reviews')}}">Reviews</a>
+                    <a class="nav-link" href="{{url_for('reviewsBp.reviews')}}">Reviews</a>
                 </li>
                 <li class="nav-item">
                     <a class="nav-link" href="{{url_for('contactoBp.contacto')}}">Contactenos</a>
