+++ Frontend/templates/base.html
@@ -46,7 +46,7 @@
             </ul>
             <ul class="navbar-nav">
                 <li class="nav-item">
-                    <a class="nav-link" href="{{url_for('reviews')}}">Reviews</a>
+                    <a class="nav-link" href="{{url_for('reviewsBp.reviews')}}">Reviews</a>
                 </li>
                 <li class="nav-item">
                     <a class="nav-link" href="{{url_for('contactoBp.contacto')}}">Contactenos</a>
