+++ Frontend/utils.py
@@ -1,0 +1,3 @@
+# Establecer aqui la ruta para hacer funcionar todas las solicitudes
+def api_URL(ruta):
+    return f'http://127.0.0.1:5000/{ruta}'
